grupo35ing2/requirements.txt
black==21.9b0
cachelib==0.7.0
click==8.1.3
colorama==0.4.4
Flask==2.0.1
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.1.2
Jinja2==3.0.1
MarkupSafe==2.1.1
mypy-extensions==0.4.3
mysqlclient==2.1.0
pathspec==0.9.0
pdfkit==1.0.0
platformdirs==2.5.2
python-dateutil==2.8.2
regex==2022.4.24
six==1.16.0
SQLAlchemy==1.4.25
SQLAlchemy-Utils==0.37.8
tomli==1.2.3
typing_extensions==4.2.0
Werkzeug==2.1.2
wkhtmltopdf==0.2
